rester/serializer.py
from rest_framework import serializers

from rester.models import User


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('username', 'password', 'mail')
