requirements.txt
Django==2.0
djangorestframework==3.7.7
Markdown==2.6.11
pytz==2017.3
